Update README.md
BIG DATA PROJECT
The goal of this project is to develop a web application based on the Apache Kafka Stream API for real-time analysis of data, with a specific focus on "predicting customer churn in real-time" for a business. Apache Kafka is a distributed event streaming platform that allows the handling of large-scale data streams efficiently.